added worker password to views

+++ ccfWorker/ccfWorker/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, reverse
 from django.views.decorators.csrf import csrf_exempt
+from django.conf import settings
 from django.http import HttpResponse
 from .invert import process
 import json
@@ -16,13 +17,15 @@ def workpage_view(request):
     '''
 
     if request.method == 'POST':
-        by = ''
-        process([os.getcwd()+'/ccfWorker/testImage.jpg'])
-        with open(os.getcwd()+'/ccfWorker/fixed.jpg',"rb") as fp:
-            by = base64.b64encode(fp.read())
-        
-        data = {'path':os.getcwd(),'url':request.COOKIES['url'],'img':str(by)}
-        os.remove(os.getcwd()+'/ccfWorker/fixed.jpg')
-        return HttpResponse(json.dumps(data))
+        if(request.COOKIES['psw'] == settings.worker_password):
+            by = ''
+            process([os.getcwd()+'/ccfWorker/testImage.jpg'])
+            with open(os.getcwd()+'/ccfWorker/fixed.jpg',"rb") as fp:
+                by = base64.b64encode(fp.read())
+            
+            data = {'path':os.getcwd(),'url':request.COOKIES['url'],'img':str(by)}
+            os.remove(os.getcwd()+'/ccfWorker/fixed.jpg')
+            return HttpResponse(json.dumps(data))
+            
     else:
         return render(request, "tmp_home.html")
